1. Added changes for display of Profile 2. Fixes for Metrics 3. Added features for User Management

+++ src/components/assumptions/personas/ProfileCard.js
@@ -7,33 +7,40 @@ import { Card, Button, Icon } from 'antd';
 class ProfileCard extends Component {
 
     componentWillReceiveProps(nextProps) {
-        this.setState({ persona: nextProps.persona });
+        this.setState({ persona: nextProps.persona, deleteDisabled: nextProps.deleteDisabled });
     }
 
     constructor(props) {
         super(props);
         this.state = {
             persona: props.persona,
-            isAvatarOpened: false
+            isAvatarOpened: false,
+            deleteDisabled: props.deleteDisabled
         };
     }
 
     render() {
+        const avatar = _.get(this.state, 'persona.avatar');
+        const fullname = _.compact([_.get(this.state, 'persona.firstName'), _.get(this.state, 'persona.lastName')]).join(" ");
         return (
             <div>
                 <Card>
                     <p>
                         <picture width="100%" onClick={this.showAvatarModal.bind(this)}>
-                          <source srcSet={ this.state.persona.avatar ? `${this.state.persona.avatar}?time=${new Date().getTime()}`: '' } />
+                          <source srcSet={ avatar ? `${this.state.persona.avatar}?time=${new Date().getTime()}`: '' } />
                           <img src="/blank-photo.jpg" alt="Default Photo" style={{borderRadius: "100%"}}/>
                         </picture>              
                     </p>
-                    <h2>{this.state.persona.firstName} {this.state.persona.lastName}</h2>
-                    <p>
-                     <Icon type="delete" onClick={this.onDelete.bind(this)} />
-                    </p>
+                    <h2>{fullname}</h2>
+                    { (this.state.deleteDisabled) ? '' :
+                        <p>
+                         <Icon type="delete" onClick={this.onDelete.bind(this)} />
+                        </p>
+                    }
                 </Card>
+                {(this.state.persona) ? 
                 <PersonaInfoForm persona={this.state.persona} ref="personaModal" beforeClose={() => this.reRender()} onUpdate={this.onUpdate.bind(this)}></PersonaInfoForm>
+                : ''}
             </div>
         );
     }
@@ -54,7 +61,7 @@ class ProfileCard extends Component {
     }
 
     onDelete() {
-        if (this.props.onDelete) {
+        if (this.props.onDelete && this.state.persona) {
             this.props.onDelete(this.state.persona);
         }
     }
